Add rename of components -> ui
