+++ packages/common/src/ui/icons/Dashboard.tsx
