+++ packages/common/src/ui/icons/MSupplyGuy.tsx
