+++ packages/common/src/ui/icons/Stock.tsx
