+++ packages/common/src/ui/icons/Suppliers.tsx
